[refactor] 리뷰에 따른 올바른 출력 수정
