+++ src/main/java/step2/domain/NumberGenerator.java
