+++ src/main/java/step2/domain/Point.java
