+++ src/main/java/step2/domain/RandomGenerator.java
